Moved legacy methods related to identify btc tx types process to PegUtilsLegacy.
